requirements.txt
Django==1.7.3
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
Pint==0.6
Unidecode==0.04.17
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
certifi==14.05.14
django-allauth==0.19.0
django-bootstrap3==5.0.3
django-chartit==0.1
django-debug-toolbar==1.2.2
django-extensions==1.5.1
django-jsonfield==0.9.13
djangorestframework==3.0.3
djorm-pgarray==1.2
fitparse==0.0.1-dev
ipython==2.3.1
oauthlib==0.7.2
psycopg2==2.5.4
pydot==1.0.2
pygraphviz==1.2
pyparsing==2.0.3
python-dateutil==1.5
python-nvd3==0.13.5
python-openid==2.2.5
python-slugify==0.0.9
pytz==2014.7
pyzmq==14.4.1
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
sqlparse==0.1.14
stravalib==0.2.2
tornado==4.0.2
units==0.6
wsgiref==0.1.2
